Encapsulate shared API logic in base API class.
This makes it easier to implement other Toggl-Based APIs sharing the same codebase, without duplication.
