+++ lib/togglv8/users.rb
@@ -1,5 +1,5 @@
 module TogglV8
-  class API
+  class API < TogglBaseApi
 
     ##
     # ---------
